packages/ps-ai-diffusion-bridge/pyproject.toml
[project]
name = "ps-ai-diffusion-bridge"
version = "0.1.0"
description = "Python Bridge service for PS AI Diffusion"
requires-python = ">=3.11"
license = {text = "GPL-3.0"}
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "pydantic>=2.0.0",
    "aiohttp>=3.9.0",
    "Pillow",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.26.0",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"
